Update advanced-threat-protection-configure.md
When integrating MDE with Intune, clicking on "Open the Microsoft Defender for Endpoint Admin Console" in the Endpoint Manager console wants to send you to securitycenter.windows.com

However it looks like when you do that, you now get redirected to the M365 defender portal. Given that, the UI is different than what is described in this article, and the path to the "Microsoft Intune Connection" setting is slightly different than what is described in this article.